Judge git repo exist by file instead of cwd
The behavior that disable blamer when current working
directory is not belong to a git root is not suitable
for some use case. User may switch between git repo
and none git repo, plugin should not be disabled forever.
In this patch, git root is detected on every buffer
enter. Once we in the git repo, we start to refresh
the blame information.  If we are not in git repo,
disable the autocmd for no overhead.